+++ irida_import/main.py
@@ -1,5 +1,7 @@
-#!/usr/bin/env python3
+from __future__ import absolute_import
+
 import argparse
+import os
 import sys
 import logging
 
@@ -23,10 +25,13 @@ if __name__ == '__main__':
         help='The tool can use a supplied access token instead of querying '
              + 'IRIDA.', metavar='token')
     parser.add_argument(
-        '-c', '--config', action='store_true', default=False, dest='config',
-        help='The tool must configure itself before Galaxy can be started. '
-             + 'Use this option to do so. config.ini should be in the main '
-             + 'irida_import folder.')
+        '-c', '--config', dest='config', default=Config.DEFAULT_CONFIG_PATH,
+        help='The tool requires a config file to run. '
+             + 'By default this is config.ini in the main irida_import folder.')
+    parser.add_argument(
+        '-g', '--generate_xml', action='store_true', default=False, dest='generate_xml',
+        help='The tool must generate a galaxy tool xml file before Galaxy can be started. '
+             + 'Use this option to do so.')
     parser.add_argument(
         '-i', '--history-id', dest='hist_id', default=False,
         help='The tool requires a History ID.')
@@ -40,27 +45,26 @@ if __name__ == '__main__':
                         format=log_format,
                         level=logging.ERROR,
                         filemode="w")
-    try:
-        config = Config()
-    except FileNotFoundError:
-        message = ('Error: Could not find config.ini in the irida_importer'
-                   + ' directory!')
+
+    if os.path.isfile(args.config):
+        config = Config(args.config)
+    else:
+        message = "Error: {} does not exist!".format(args.config)
         logging.info(message)
-        print(message)
-        exit(1)
+        sys.exit(message)
 
     logging.debug("Reading from passed file")
 
-    if args.config:
-        config.emit_tool_xml()
-        message = 'Successfully configured the XML file!'
+    if args.generate_xml:
+        config.generate_xml()
+        message = 'Successfully generated the XML file!'
         logging.info(message)
         print(message)
     else:
         # importing here prevents a user from needing all libs when only performing a config,
         # after the tool xml is generated galaxy will install all the required dependencies
         from irida_import.irida_import import IridaImport
-        
+
         importer = IridaImport(config)
         # otherwise start looking at the input file
         try: