+++ irida_import/tests/integration/bash_scripts/install_galaxy.sh
@@ -4,26 +4,7 @@ args=("$@")
 tool_loc=${args[0]}
 galaxy_port=${args[1]}
 
-mkdir repos
-pushd repos
-
-echo "Downloading IRIDA..."
-git clone https://github.com/phac-nml/irida.git
-pushd irida
-git checkout master > irida-checkout.log 2>&1
-git fetch
-git reset --hard
-git clean -fd
-git pull
-echo "Preparing IRIDA for first excecution..."
-rm -rf /tmp/shed_tools/
-pkill -u $USER -f "python ./scripts/paster.py" || true
-
-pushd lib
-./install-libs.sh > /dev/null
-popd
-popd
-echo "IRIDA has been installed"
+pushd /tmp/repos
 
 echo "Downloading Galaxy..."
 git clone https://github.com/galaxyproject/galaxy/ > galaxy-clone.log 2>&1
@@ -35,6 +16,10 @@ git clean -fd
 git pull
 echo "Preparing Galaxy for first execution (installing eggs)..."
 ./scripts/common_startup.sh > galaxy-common-startup.log 2>&1
+
+# reset galaxy psql db
+echo "dropping galaxy_test database in psql"
+echo "drop database if exists galaxy_test; create database galaxy_test;" | psql
 
 echo "Configuring Galaxy."
 pushd config
@@ -55,7 +40,8 @@ echo "  database_connection: postgresql:///galaxy_test" | cat >> galaxy.yml
 sed -i 's/#admin_users:.*/admin_users: "irida@irida.ca"/' galaxy.yml
 
 # run galaxy on port 8888 instead of 8080; Tomcat runs on 8080 by default.
-sed -i "s|http: 127.0.0.1:8080|http: 127.0.0.1:$galaxy_port|" galaxy.yml
+sed -i "s|# bind: localhost:8080|bind: localhost:$galaxy_port|" galaxy.yml
+
 popd
 popd
 echo "Galaxy has been installed"
@@ -82,4 +68,4 @@ popd
 
 popd
 
-echo "IRIDA, Galaxy, and the IRIDA export tool have been installed"
+echo "Galaxy and the IRIDA export tool have been installed"