+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 All notable changes to irida-galaxy-importer will be documeted in this file.
 
+## 2.0.1
+* Switched from travisCI tests to Github Actions
+
 ## 2.0.0
 
 * Added support for Python 3, while maintaining Python 2 support